+++ history.md
@@ -4,6 +4,7 @@
 * When working with missing generic arguments for embedded resources it assumed that all embedded resource are at least HyperResource interfaces, beyond that they also will implement any interfaces as defined by the resource interface selection strategy.
  * In other words supposed you have a Page<T> with a T[] items() method.  You can now have code like ```Page menu = (Page)resource.thatHasSomethingReturningHyperResource();  Object[] items = menu.items();  if(items[0] instanceof Resourcetype)``` to get out of the generic Object and back into specialized interfaces.
  * Very useful when you can't be sure what is being returned and send it through a view bindinging routing based on the resource you have
+* You can now remove request interceptors by type
 
 ## 1.9.0 - 2015-12-16
 * Fixed a bug where the Response object of okhttp2 client only included the last header with a given name